Renamed template folder to template_rendering
